+++ test/ClangImporter/Inputs/custom-modules/UnimportableMembersUser.h
@@ -1,5 +1,5 @@
 @import UnimportableMembers;
 
 @interface DesignatedInitializerInAnotherModule (/*evil class extension*/)
-- (instancetype)initFromOtherModule:(long)x __attribute__((objc_designated_initializer));
+- (instancetype)initFromOtherModule:(intptr_t)x __attribute__((objc_designated_initializer));
 @end
